chore(richtext-lexical): split up feature types in client & server feature types
