+++ packages/richtext-lexical/src/features/blocks/component/index.tsx
@@ -25,7 +25,7 @@ import { getTranslation } from '@payloadcms/translations'
 import { getFormState } from '@payloadcms/ui/shared'
 import { v4 as uuid } from 'uuid'
 
-import type { ClientComponentProps } from '../../types.js'
+import type { ClientComponentProps } from '../../typesClient.js'
 import type { BlocksFeatureClientProps } from '../feature.client.js'
 
 import { useEditorConfigContext } from '../../../lexical/config/client/EditorConfigProvider.js'
