+++ packages/richtext-lexical/src/features/createClientComponent.tsx
@@ -1,6 +1,7 @@
 'use client'
 
-import type { FeatureProviderProviderClient, ServerFeature } from './types.js'
+import type { FeatureProviderProviderClient } from './typesClient.js'
+import type { ServerFeature } from './typesServer.js'
 
 import { useLexicalFeature } from '../utilities/useLexicalFeature.js'
