+++ packages/richtext-lexical/src/features/link/plugins/autoLink/index.tsx
@@ -16,7 +16,7 @@ import {
 } from 'lexical'
 import { useEffect } from 'react'
 
-import type { PluginComponent } from '../../../types.js'
+import type { PluginComponent } from '../../../typesClient.js'
 import type { ClientProps } from '../../feature.client.js'
 import type { LinkFields } from '../../nodes/types.js'
