+++ packages/richtext-lexical/src/features/link/plugins/clickableLink/index.tsx
@@ -2,7 +2,7 @@
 import { ClickableLinkPlugin as LexicalClickableLinkPlugin } from '@lexical/react/LexicalClickableLinkPlugin.js'
 import React from 'react'
 
-import type { PluginComponent } from '../../../types.js'
+import type { PluginComponent } from '../../../typesClient.js'
 import type { ClientProps } from '../../feature.client.js'
 
 export const ClickableLinkPlugin: PluginComponent<ClientProps> = () => {
