+++ packages/richtext-lexical/src/features/link/validate.ts
@@ -2,7 +2,7 @@ import type { Field } from 'payload'
 
 import { buildStateFromSchema } from '@payloadcms/ui/forms/buildStateFromSchema'
 
-import type { NodeValidation } from '../types.js'
+import type { NodeValidation } from '../typesServer.js'
 import type { LinkFeatureServerProps } from './feature.server.js'
 import type { SerializedAutoLinkNode, SerializedLinkNode } from './nodes/types.js'
