+++ packages/richtext-lexical/src/features/lists/checklist/plugin/index.tsx
@@ -2,7 +2,7 @@
 import { CheckListPlugin } from '@lexical/react/LexicalCheckListPlugin.js'
 import React from 'react'
 
-import type { PluginComponent } from '../../../types.js'
+import type { PluginComponent } from '../../../typesClient.js'
 
 export const LexicalCheckListPlugin: PluginComponent<undefined> = () => {
   return <CheckListPlugin />
