+++ packages/richtext-lexical/src/features/lists/plugin/index.tsx
@@ -2,7 +2,7 @@
 import { ListPlugin } from '@lexical/react/LexicalListPlugin.js'
 import React from 'react'
 
-import type { PluginComponent } from '../../types.js'
+import type { PluginComponent } from '../../typesClient.js'
 
 export const LexicalListPlugin: PluginComponent<undefined> = () => {
   return <ListPlugin />
