+++ packages/richtext-lexical/src/features/toolbars/fixed/feature.client.tsx
@@ -1,26 +1,15 @@
 'use client'
 
-import type { FeatureProviderProviderClient } from '../../types.js'
 import type { FixedToolbarFeatureProps } from './feature.server.js'
 
-import { createClientComponent } from '../../createClientComponent.js'
+import { createClientFeature } from '../../../utilities/createClientFeature.js'
 import { FixedToolbarPlugin } from './Toolbar/index.js'
 
-const FixedToolbarFeatureClient: FeatureProviderProviderClient<FixedToolbarFeatureProps> = (
-  props,
-) => {
-  return {
-    clientFeatureProps: props,
-    feature: {
-      plugins: [
-        {
-          Component: FixedToolbarPlugin,
-          position: 'aboveContainer',
-        },
-      ],
-      sanitizedClientFeatureProps: props,
+export const FixedToolbarFeatureClient = createClientFeature<FixedToolbarFeatureProps>({
+  plugins: [
+    {
+      Component: FixedToolbarPlugin,
+      position: 'aboveContainer',
     },
-  }
-}
-
-export const FixedToolbarFeatureClientComponent = createClientComponent(FixedToolbarFeatureClient)
+  ],
+})
