+++ packages/richtext-lexical/src/features/toolbars/inline/feature.client.tsx
@@ -1,23 +1,13 @@
 'use client'
 
-import type { FeatureProviderProviderClient } from '../../types.js'
-
-import { createClientComponent } from '../../createClientComponent.js'
+import { createClientFeature } from '../../../utilities/createClientFeature.js'
 import { InlineToolbarPlugin } from './Toolbar/index.js'
 
-const InlineToolbarFeatureClient: FeatureProviderProviderClient<undefined> = (props) => {
-  return {
-    clientFeatureProps: props,
-    feature: () => ({
-      plugins: [
-        {
-          Component: InlineToolbarPlugin,
-          position: 'floatingAnchorElem',
-        },
-      ],
-      sanitizedClientFeatureProps: props,
-    }),
-  }
-}
-
-export const InlineToolbarFeatureClientComponent = createClientComponent(InlineToolbarFeatureClient)
+export const InlineToolbarFeatureClient = createClientFeature({
+  plugins: [
+    {
+      Component: InlineToolbarPlugin,
+      position: 'floatingAnchorElem',
+    },
+  ],
+})
