+++ packages/richtext-lexical/src/features/typeUtilities.ts
@@ -1,6 +1,6 @@
 import type { LexicalNode } from 'lexical'
 
-import type { NodeWithHooks } from './types.js'
+import type { NodeWithHooks } from './typesServer.js'
 
 /**
  * Utility function to create a node with hooks. You don't have to use this utility, but it improves type inference
