+++ packages/richtext-lexical/src/features/typesServer.ts
@@ -3,7 +3,6 @@ import type { GenericLanguages, I18nClient } from '@payloadcms/translations'
 import type { JSONSchema4 } from 'json-schema'
 import type {
   Klass,
-  LexicalEditor,
   LexicalNode,
   LexicalNodeReplacement,
   SerializedEditorState,
@@ -20,11 +19,10 @@ import type {
 } from 'payload'
 import type React from 'react'
 
-import type { ClientEditorConfig, ServerEditorConfig } from '../lexical/config/types.js'
-import type { SlashMenuGroup } from '../lexical/plugins/SlashMenu/LexicalTypeaheadMenuPlugin/types.js'
+import type { ServerEditorConfig } from '../lexical/config/types.js'
 import type { AdapterProps } from '../types.js'
 import type { HTMLConverter } from './converters/html/converter/types.js'
-import type { ToolbarGroup } from './toolbars/types.js'
+import type { ClientComponentProps } from './typesClient.js'
 
 export type PopulationPromise<T extends SerializedLexicalNode = SerializedLexicalNode> = ({
   context,
@@ -92,11 +90,11 @@ export type FeatureProviderServer<
   ServerFeatureProps = UnSanitizedServerFeatureProps,
   ClientFeatureProps = undefined,
 > = {
-  /** Keys of dependencies needed for this feature. These dependencies do not have to be loaded first, but they have to exist, otherwise an error will be thrown. */
+  /** Keys of dependencies needed for this feature. These dependencies do not have to be loaded first */
   dependencies?: string[]
-  /**  Keys of priority dependencies needed for this feature. These dependencies have to be loaded first AND have to exist, otherwise an error will be thrown. They will be available in the `feature` property. */
+  /** Keys of priority dependencies needed for this feature. These dependencies have to be loaded first and are available in the `feature` property*/
   dependenciesPriority?: string[]
-  /** Keys of soft-dependencies needed for this feature. These are optional. Payload will attempt to load them before this feature, but doesn't throw an error if that's not possible. */
+  /** Keys of soft-dependencies needed for this feature. The FeatureProviders dependencies are optional, but are considered as last-priority in the loading process */
   dependenciesSoft?: string[]
 
   /**
@@ -120,143 +118,6 @@ export type FeatureProviderServer<
   /** Props which were passed into your feature will have to be passed here. This will allow them to be used / read in other places of the code, e.g. wherever you can use useEditorConfigContext */
   serverFeatureProps: UnSanitizedServerFeatureProps
 }
-
-export type FeatureProviderProviderClient<
-  UnSanitizedClientFeatureProps = undefined,
-  ClientFeatureProps = UnSanitizedClientFeatureProps,
-> = (props: ClientComponentProps<ClientFeatureProps>) => FeatureProviderClient<ClientFeatureProps>
-
-/**
- * No dependencies => Features need to be sorted on the server first, then sent to client in right order
- */
-export type FeatureProviderClient<
-  UnSanitizedClientFeatureProps = undefined,
-  ClientFeatureProps = UnSanitizedClientFeatureProps,
-> = {
-  /**
-   * Return props, to make it easy to retrieve passed in props to this Feature for the client if anyone wants to
-   */
-  clientFeatureProps: ClientComponentProps<UnSanitizedClientFeatureProps>
-  feature:
-    | ((props: {
-        clientFunctions: Record<string, any>
-        /** unSanitizedEditorConfig.features, but mapped */
-        featureProviderMap: ClientFeatureProviderMap
-        // other resolved features, which have been loaded before this one. All features declared in 'dependencies' should be available here
-        resolvedFeatures: ResolvedClientFeatureMap
-        // unSanitized EditorConfig,
-        unSanitizedEditorConfig: ClientEditorConfig
-      }) => ClientFeature<ClientFeatureProps>)
-    | ClientFeature<ClientFeatureProps>
-}
-
-export type PluginComponent<ClientFeatureProps = any> = React.FC<{
-  clientProps: ClientFeatureProps
-}>
-export type PluginComponentWithAnchor<ClientFeatureProps = any> = React.FC<{
-  anchorElem: HTMLElement
-  clientProps: ClientFeatureProps
-}>
-
-export type ClientFeature<ClientFeatureProps> = {
-  hooks?: {
-    load?: ({
-      incomingEditorState,
-    }: {
-      incomingEditorState: SerializedEditorState
-    }) => SerializedEditorState
-    save?: ({
-      incomingEditorState,
-    }: {
-      incomingEditorState: SerializedEditorState
-    }) => SerializedEditorState
-  }
-  markdownTransformers?: Transformer[]
-  nodes?: Array<Klass<LexicalNode> | LexicalNodeReplacement>
-  /**
-   * Plugins are react components which get added to the editor. You can use them to interact with lexical, e.g. to create a command which creates a node, or opens a modal, or some other more "outside" functionality
-   */
-  plugins?: Array<
-    | {
-        // plugins are anything which is not directly part of the editor. Like, creating a command which creates a node, or opens a modal, or some other more "outside" functionality
-        Component: PluginComponent<ClientFeatureProps>
-        position: 'aboveContainer' // Determines at which position the Component will be added.
-      }
-    | {
-        // plugins are anything which is not directly part of the editor. Like, creating a command which creates a node, or opens a modal, or some other more "outside" functionality
-        Component: PluginComponent<ClientFeatureProps>
-        position: 'bottom' // Determines at which position the Component will be added.
-      }
-    | {
-        // plugins are anything which is not directly part of the editor. Like, creating a command which creates a node, or opens a modal, or some other more "outside" functionality
-        Component: PluginComponent<ClientFeatureProps>
-        position: 'normal' // Determines at which position the Component will be added.
-      }
-    | {
-        // plugins are anything which is not directly part of the editor. Like, creating a command which creates a node, or opens a modal, or some other more "outside" functionality
-        Component: PluginComponent<ClientFeatureProps>
-        position: 'top' // Determines at which position the Component will be added.
-      }
-    | {
-        // plugins are anything which is not directly part of the editor. Like, creating a command which creates a node, or opens a modal, or some other more "outside" functionality
-        Component: PluginComponentWithAnchor<ClientFeatureProps>
-        position: 'floatingAnchorElem' // Determines at which position the Component will be added.
-      }
-    | {
-        Component: PluginComponent<ClientFeatureProps>
-        position: 'belowContainer' // Determines at which position the Component will be added.
-      }
-  >
-  /**
-   * Return props, to make it easy to retrieve passed in props to this Feature for the client if anyone wants to
-   */
-  sanitizedClientFeatureProps?: ClientComponentProps<ClientFeatureProps>
-  slashMenu?: {
-    /**
-     * Dynamic groups allow you to add different groups depending on the query string (so, the text after the slash).
-     * Thus, to re-calculate the available groups, this function will be called every time you type after the /.
-     *
-     * The groups provided by dynamicGroups will be merged with the static groups provided by the groups property.
-     */
-    dynamicGroups?: ({
-      editor,
-      queryString,
-    }: {
-      editor: LexicalEditor
-      queryString: string
-    }) => SlashMenuGroup[]
-    /**
-     * Static array of groups together with the items in them. These will always be present.
-     * While typing after the /, they will be filtered by the query string and the keywords, key and display name of the items.
-     */
-    groups?: SlashMenuGroup[]
-  }
-  /**
-   * An opt-in, classic fixed toolbar which stays at the top of the editor
-   */
-  toolbarFixed?: {
-    groups: ToolbarGroup[]
-  }
-  /**
-   * The default, floating toolbar which appears when you select text.
-   */
-  toolbarInline?: {
-    /**
-     * Array of toolbar groups / sections. Each section can contain multiple toolbar items.
-     */
-    groups: ToolbarGroup[]
-  }
-}
-
-export type ClientComponentProps<ClientFeatureProps> = ClientFeatureProps extends undefined
-  ? {
-      featureKey: string
-      order: number
-    }
-  : {
-      featureKey: string
-      order: number
-    } & ClientFeatureProps
 
 export type AfterReadNodeHookArgs<T extends SerializedLexicalNode> = {
   /**
@@ -472,62 +333,9 @@ export type ResolvedServerFeature<ServerProps, ClientFeatureProps> = ServerFeatu
     order: number
   }
 
-export type ResolvedClientFeature<ClientFeatureProps> = ClientFeature<ClientFeatureProps> & {
-  key: string
-  order: number
-}
-
 export type ResolvedServerFeatureMap = Map<string, ResolvedServerFeature<unknown, unknown>>
-export type ResolvedClientFeatureMap = Map<string, ResolvedClientFeature<unknown>>
 
 export type ServerFeatureProviderMap = Map<string, FeatureProviderServer<unknown, unknown, unknown>>
-export type ClientFeatureProviderMap = Map<string, FeatureProviderClient<unknown, unknown>>
-
-/**
- * Plugins are react components which get added to the editor. You can use them to interact with lexical, e.g. to create a command which creates a node, or opens a modal, or some other more "outside" functionality
- */
-export type SanitizedPlugin =
-  | {
-      // plugins are anything which is not directly part of the editor. Like, creating a command which creates a node, or opens a modal, or some other more "outside" functionality
-      Component: PluginComponent
-      clientProps: any
-      key: string
-      position: 'bottom' // Determines at which position the Component will be added.
-    }
-  | {
-      // plugins are anything which is not directly part of the editor. Like, creating a command which creates a node, or opens a modal, or some other more "outside" functionality
-      Component: PluginComponent
-      clientProps: any
-      key: string
-      position: 'normal' // Determines at which position the Component will be added.
-    }
-  | {
-      // plugins are anything which is not directly part of the editor. Like, creating a command which creates a node, or opens a modal, or some other more "outside" functionality
-      Component: PluginComponent
-      clientProps: any
-      key: string
-      position: 'top' // Determines at which position the Component will be added.
-    }
-  | {
-      // plugins are anything which is not directly part of the editor. Like, creating a command which creates a node, or opens a modal, or some other more "outside" functionality
-      Component: PluginComponentWithAnchor
-      clientProps: any
-      desktopOnly?: boolean
-      key: string
-      position: 'floatingAnchorElem' // Determines at which position the Component will be added.
-    }
-  | {
-      Component: PluginComponent
-      clientProps: any
-      key: string
-      position: 'aboveContainer'
-    }
-  | {
-      Component: PluginComponent
-      clientProps: any
-      key: string
-      position: 'belowContainer'
-    }
 
 export type SanitizedServerFeatures = Required<
   Pick<ResolvedServerFeature<unknown, unknown>, 'i18n' | 'markdownTransformers' | 'nodes'>
@@ -582,50 +390,4 @@ export type SanitizedServerFeatures = Required<
   } /**  The node types mapped to their populationPromises */
   /**  The node types mapped to their validations */
   validations: Map<string, Array<NodeValidation>>
-}
-
-export type SanitizedClientFeatures = Required<
-  Pick<
-    ResolvedClientFeature<unknown>,
-    'markdownTransformers' | 'nodes' | 'toolbarFixed' | 'toolbarInline'
-  >
-> & {
-  /** The keys of all enabled features */
-  enabledFeatures: string[]
-  hooks: {
-    load: Array<
-      ({
-        incomingEditorState,
-      }: {
-        incomingEditorState: SerializedEditorState
-      }) => SerializedEditorState
-    >
-    save: Array<
-      ({
-        incomingEditorState,
-      }: {
-        incomingEditorState: SerializedEditorState
-      }) => SerializedEditorState
-    >
-  }
-  /**
-   * Plugins are react components which get added to the editor. You can use them to interact with lexical, e.g. to create a command which creates a node, or opens a modal, or some other more "outside" functionality
-   */
-  plugins?: Array<SanitizedPlugin>
-  slashMenu: {
-    /**
-     * Dynamic groups allow you to add different groups depending on the query string (so, the text after the slash).
-     * Thus, to re-calculate the available groups, this function will be called every time you type after the /.
-     *
-     * The groups provided by dynamicGroups will be merged with the static groups provided by the groups property.
-     */
-    dynamicGroups: Array<
-      ({ editor, queryString }: { editor: LexicalEditor; queryString: string }) => SlashMenuGroup[]
-    >
-    /**
-     * Static array of groups together with the items in them. These will always be present.
-     * While typing after the /, they will be filtered by the query string and the keywords, key and display name of the items.
-     */
-    groups: SlashMenuGroup[]
-  }
 }
