+++ packages/richtext-lexical/src/features/upload/feature.server.ts
@@ -11,12 +11,12 @@ import type {
 import { traverseFields } from '@payloadcms/ui/utilities/buildFieldSchemaMap/traverseFields'
 import { sanitizeFields } from 'payload'
 
-import type { FeatureProviderProviderServer } from '../types.js'
 import type { UploadFeaturePropsClient } from './feature.client.js'
 
 // eslint-disable-next-line payload/no-imports-from-exports-dir
-import { UploadFeatureClientComponent } from '../../exports/client/index.js'
+import { UploadFeatureClient } from '../../exports/client/index.js'
 import { populate } from '../../populateGraphQL/populate.js'
+import { createServerFeature } from '../../utilities/createServerFeature.js'
 import { createNode } from '../typeUtilities.js'
 import { uploadPopulationPromiseHOC } from './graphQLPopulationPromise.js'
 import { i18n } from './i18n.js'
@@ -45,235 +45,230 @@ function getAbsoluteURL(url: string, payload: Payload): string {
   return url?.startsWith('http') ? url : (payload?.config?.serverURL || '') + url
 }
 
-export const UploadFeature: FeatureProviderProviderServer<
+export const UploadFeature = createServerFeature<
   UploadFeatureProps,
   UploadFeatureProps,
   UploadFeaturePropsClient
-> = (props) => {
-  if (!props) {
-    props = { collections: {} }
-  }
-
-  const clientProps: UploadFeaturePropsClient = {
-    collections: {},
-  }
-  if (props.collections) {
-    for (const collection in props.collections) {
-      clientProps.collections[collection] = {
-        hasExtraFields: props.collections[collection].fields.length >= 1,
-      }
+>({
+  feature: async ({ config: _config, isRoot, props }) => {
+    if (!props) {
+      props = { collections: {} }
     }
-  }
 
-  return {
-    feature: async ({ config: _config, isRoot }) => {
-      const validRelationships = _config.collections.map((c) => c.slug) || []
-
+    const clientProps: UploadFeaturePropsClient = {
+      collections: {},
+    }
+    if (props.collections) {
       for (const collection in props.collections) {
-        if (props.collections[collection].fields?.length) {
-          props.collections[collection].fields = await sanitizeFields({
-            config: _config as unknown as Config,
-            fields: props.collections[collection].fields,
-            requireFieldLevelRichTextEditor: isRoot,
-            validRelationships,
-          })
+        clientProps.collections[collection] = {
+          hasExtraFields: props.collections[collection].fields.length >= 1,
         }
       }
+    }
 
-      return {
-        ClientFeature: UploadFeatureClientComponent,
-        clientFeatureProps: clientProps,
-        generateSchemaMap: ({ config, i18n, props }) => {
-          if (!props?.collections) return null
+    const validRelationships = _config.collections.map((c) => c.slug) || []
 
-          const schemaMap = new Map<string, Field[]>()
-          const validRelationships = config.collections.map((c) => c.slug) || []
-
-          for (const collection in props.collections) {
-            if (props.collections[collection].fields?.length) {
-              schemaMap.set(collection, props.collections[collection].fields)
-              traverseFields({
-                config,
-                fields: props.collections[collection].fields,
-                i18n,
-                schemaMap,
-                schemaPath: collection,
-                validRelationships,
-              })
-            }
-          }
-
-          return schemaMap
-        },
-        i18n,
-        nodes: [
-          createNode({
-            converters: {
-              html: {
-                converter: async ({ node, req }) => {
-                  // @ts-expect-error
-                  const id = node?.value?.id || node?.value // for backwards-compatibility
-
-                  if (req?.payload) {
-                    const uploadDocument: {
-                      value?: TypeWithID & FileData
-                    } = {}
-
-                    try {
-                      await populate({
-                        id,
-                        collectionSlug: node.relationTo,
-                        currentDepth: 0,
-                        data: uploadDocument,
-                        depth: 1,
-                        draft: false,
-                        key: 'value',
-                        overrideAccess: false,
-                        req,
-                        showHiddenFields: false,
-                      })
-                    } catch (ignored) {
-                      // eslint-disable-next-line no-console
-                      console.error(
-                        'Lexical upload node HTML converter: error fetching upload file',
-                        ignored,
-                        'Node:',
-                        node,
-                      )
-                      return `<img />`
-                    }
-
-                    const url = getAbsoluteURL(uploadDocument?.value?.url, req?.payload)
-
-                    /**
-                     * If the upload is not an image, return a link to the upload
-                     */
-                    if (!uploadDocument?.value?.mimeType?.startsWith('image')) {
-                      return `<a href="${url}" rel="noopener noreferrer">${uploadDocument.value?.filename}</a>`
-                    }
-
-                    /**
-                     * If the upload is a simple image with no different sizes, return a simple img tag
-                     */
-                    if (
-                      !uploadDocument?.value?.sizes ||
-                      !Object.keys(uploadDocument?.value?.sizes).length
-                    ) {
-                      return `<img src="${url}" alt="${uploadDocument?.value?.filename}" width="${uploadDocument?.value?.width}"  height="${uploadDocument?.value?.height}"/>`
-                    }
-
-                    /**
-                     * If the upload is an image with different sizes, return a picture element
-                     */
-                    let pictureHTML = '<picture>'
-
-                    // Iterate through each size in the data.sizes object
-                    for (const size in uploadDocument.value?.sizes) {
-                      const imageSize: FileSize & {
-                        url?: string
-                      } = uploadDocument.value?.sizes[size]
-
-                      // Skip if any property of the size object is null
-                      if (
-                        !imageSize.width ||
-                        !imageSize.height ||
-                        !imageSize.mimeType ||
-                        !imageSize.filesize ||
-                        !imageSize.filename ||
-                        !imageSize.url
-                      ) {
-                        continue
-                      }
-                      const imageSizeURL = getAbsoluteURL(imageSize?.url, req?.payload)
-
-                      pictureHTML += `<source srcset="${imageSizeURL}" media="(max-width: ${imageSize.width}px)" type="${imageSize.mimeType}">`
-                    }
-
-                    // Add the default img tag
-                    pictureHTML += `<img src="${url}" alt="Image" width="${uploadDocument.value?.width}" height="${uploadDocument.value?.height}">`
-                    pictureHTML += '</picture>'
-                    return pictureHTML
-                  } else {
-                    return `<img src="${id}" />`
-                  }
-                },
-                nodeTypes: [UploadNode.getType()],
-              },
-            },
-            getSubFields: ({ node, req }) => {
-              const collection = req.payload.collections[node?.relationTo]
-
-              if (collection) {
-                const collectionFieldSchema = props?.collections?.[node?.relationTo]?.fields
-
-                if (Array.isArray(collectionFieldSchema)) {
-                  if (!collectionFieldSchema?.length) {
-                    return null
-                  }
-                  return collectionFieldSchema
-                }
-              }
-              return null
-            },
-            getSubFieldsData: ({ node }) => {
-              return node?.fields
-            },
-            graphQLPopulationPromises: [uploadPopulationPromiseHOC(props)],
-            hooks: {
-              afterRead: [
-                ({
-                  currentDepth,
-                  depth,
-                  draft,
-                  node,
-                  overrideAccess,
-                  populationPromises,
-                  req,
-                  showHiddenFields,
-                }) => {
-                  if (!node?.value) {
-                    return node
-                  }
-                  const collection = req.payload.collections[node?.relationTo]
-
-                  if (!collection) {
-                    return node
-                  }
-                  // @ts-expect-error
-                  const id = node?.value?.id || node?.value // for backwards-compatibility
-
-                  const populateDepth =
-                    props?.maxDepth !== undefined && props?.maxDepth < depth
-                      ? props?.maxDepth
-                      : depth
-
-                  populationPromises.push(
-                    populate({
-                      id,
-                      collectionSlug: collection.config.slug,
-                      currentDepth,
-                      data: node,
-                      depth: populateDepth,
-                      draft,
-                      key: 'value',
-                      overrideAccess,
-                      req,
-                      showHiddenFields,
-                    }),
-                  )
-
-                  return node
-                },
-              ],
-            },
-            node: UploadNode,
-            validations: [uploadValidation(props)],
-          }),
-        ],
-        sanitizedServerFeatureProps: props,
+    for (const collection in props.collections) {
+      if (props.collections[collection].fields?.length) {
+        props.collections[collection].fields = await sanitizeFields({
+          config: _config as unknown as Config,
+          fields: props.collections[collection].fields,
+          requireFieldLevelRichTextEditor: isRoot,
+          validRelationships,
+        })
       }
-    },
-    key: 'upload',
-    serverFeatureProps: props,
-  }
-}
+    }
+
+    return {
+      ClientFeature: UploadFeatureClient,
+      clientFeatureProps: clientProps,
+      generateSchemaMap: ({ config, i18n, props }) => {
+        if (!props?.collections) return null
+
+        const schemaMap = new Map<string, Field[]>()
+        const validRelationships = config.collections.map((c) => c.slug) || []
+
+        for (const collection in props.collections) {
+          if (props.collections[collection].fields?.length) {
+            schemaMap.set(collection, props.collections[collection].fields)
+            traverseFields({
+              config,
+              fields: props.collections[collection].fields,
+              i18n,
+              schemaMap,
+              schemaPath: collection,
+              validRelationships,
+            })
+          }
+        }
+
+        return schemaMap
+      },
+      i18n,
+      nodes: [
+        createNode({
+          converters: {
+            html: {
+              converter: async ({ node, req }) => {
+                // @ts-expect-error
+                const id = node?.value?.id || node?.value // for backwards-compatibility
+
+                if (req?.payload) {
+                  const uploadDocument: {
+                    value?: TypeWithID & FileData
+                  } = {}
+
+                  try {
+                    await populate({
+                      id,
+                      collectionSlug: node.relationTo,
+                      currentDepth: 0,
+                      data: uploadDocument,
+                      depth: 1,
+                      draft: false,
+                      key: 'value',
+                      overrideAccess: false,
+                      req,
+                      showHiddenFields: false,
+                    })
+                  } catch (ignored) {
+                    // eslint-disable-next-line no-console
+                    console.error(
+                      'Lexical upload node HTML converter: error fetching upload file',
+                      ignored,
+                      'Node:',
+                      node,
+                    )
+                    return `<img />`
+                  }
+
+                  const url = getAbsoluteURL(uploadDocument?.value?.url, req?.payload)
+
+                  /**
+                   * If the upload is not an image, return a link to the upload
+                   */
+                  if (!uploadDocument?.value?.mimeType?.startsWith('image')) {
+                    return `<a href="${url}" rel="noopener noreferrer">${uploadDocument.value?.filename}</a>`
+                  }
+
+                  /**
+                   * If the upload is a simple image with no different sizes, return a simple img tag
+                   */
+                  if (
+                    !uploadDocument?.value?.sizes ||
+                    !Object.keys(uploadDocument?.value?.sizes).length
+                  ) {
+                    return `<img src="${url}" alt="${uploadDocument?.value?.filename}" width="${uploadDocument?.value?.width}"  height="${uploadDocument?.value?.height}"/>`
+                  }
+
+                  /**
+                   * If the upload is an image with different sizes, return a picture element
+                   */
+                  let pictureHTML = '<picture>'
+
+                  // Iterate through each size in the data.sizes object
+                  for (const size in uploadDocument.value?.sizes) {
+                    const imageSize: FileSize & {
+                      url?: string
+                    } = uploadDocument.value?.sizes[size]
+
+                    // Skip if any property of the size object is null
+                    if (
+                      !imageSize.width ||
+                      !imageSize.height ||
+                      !imageSize.mimeType ||
+                      !imageSize.filesize ||
+                      !imageSize.filename ||
+                      !imageSize.url
+                    ) {
+                      continue
+                    }
+                    const imageSizeURL = getAbsoluteURL(imageSize?.url, req?.payload)
+
+                    pictureHTML += `<source srcset="${imageSizeURL}" media="(max-width: ${imageSize.width}px)" type="${imageSize.mimeType}">`
+                  }
+
+                  // Add the default img tag
+                  pictureHTML += `<img src="${url}" alt="Image" width="${uploadDocument.value?.width}" height="${uploadDocument.value?.height}">`
+                  pictureHTML += '</picture>'
+                  return pictureHTML
+                } else {
+                  return `<img src="${id}" />`
+                }
+              },
+              nodeTypes: [UploadNode.getType()],
+            },
+          },
+          getSubFields: ({ node, req }) => {
+            const collection = req.payload.collections[node?.relationTo]
+
+            if (collection) {
+              const collectionFieldSchema = props?.collections?.[node?.relationTo]?.fields
+
+              if (Array.isArray(collectionFieldSchema)) {
+                if (!collectionFieldSchema?.length) {
+                  return null
+                }
+                return collectionFieldSchema
+              }
+            }
+            return null
+          },
+          getSubFieldsData: ({ node }) => {
+            return node?.fields
+          },
+          graphQLPopulationPromises: [uploadPopulationPromiseHOC(props)],
+          hooks: {
+            afterRead: [
+              ({
+                currentDepth,
+                depth,
+                draft,
+                node,
+                overrideAccess,
+                populationPromises,
+                req,
+                showHiddenFields,
+              }) => {
+                if (!node?.value) {
+                  return node
+                }
+                const collection = req.payload.collections[node?.relationTo]
+
+                if (!collection) {
+                  return node
+                }
+                // @ts-expect-error
+                const id = node?.value?.id || node?.value // for backwards-compatibility
+
+                const populateDepth =
+                  props?.maxDepth !== undefined && props?.maxDepth < depth ? props?.maxDepth : depth
+
+                populationPromises.push(
+                  populate({
+                    id,
+                    collectionSlug: collection.config.slug,
+                    currentDepth,
+                    data: node,
+                    depth: populateDepth,
+                    draft,
+                    key: 'value',
+                    overrideAccess,
+                    req,
+                    showHiddenFields,
+                  }),
+                )
+
+                return node
+              },
+            ],
+          },
+          node: UploadNode,
+          validations: [uploadValidation(props)],
+        }),
+      ],
+      sanitizedServerFeatureProps: props,
+    }
+  },
+  key: 'upload',
+})
