+++ packages/richtext-lexical/src/lexical/config/client/sanitize.ts
@@ -2,7 +2,10 @@
 
 import type { EditorConfig as LexicalEditorConfig } from 'lexical'
 
-import type { ResolvedClientFeatureMap, SanitizedClientFeatures } from '../../../features/types.js'
+import type {
+  ResolvedClientFeatureMap,
+  SanitizedClientFeatures,
+} from '../../../features/typesClient.js'
 import type { LexicalFieldAdminProps } from '../../../types.js'
 import type { SanitizedClientEditorConfig } from '../types.js'
