+++ packages/richtext-lexical/src/lexical/config/server/default.ts
@@ -1,6 +1,6 @@
 import type { EditorConfig as LexicalEditorConfig } from 'lexical'
 
-import type { FeatureProviderServer } from '../../../features/types.js'
+import type { FeatureProviderServer } from '../../../features/typesServer.js'
 import type { ServerEditorConfig } from '../types.js'
 
 import { AlignFeature } from '../../../features/align/feature.server.js'
