+++ packages/richtext-lexical/src/populateGraphQL/populateLexicalPopulationPromises.ts
@@ -1,7 +1,7 @@
 import type { SerializedEditorState } from 'lexical'
 import type { RichTextAdapter } from 'payload'
 
-import type { PopulationPromise } from '../features/types.js'
+import type { PopulationPromise } from '../features/typesServer.js'
 import type { AdapterProps } from '../types.js'
 
 import { recurseNodes } from '../utilities/forEachNodeRecursively.js'
